pyproject.toml
[project]
name = "pcf-pipeline"
version = "0.1.0"
description = "Phenocycler TMA data processing pipeline with Nimbus inference"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "tifffile",
    "imagecodecs",
    "scikit-image",
    "anndata",
    "pyyaml",
    "pyarrow",
]

[project.optional-dependencies]
nimbus = ["nimbus-inference"]
dev = ["ruff", "pytest"]

[project.scripts]
pcf-pipeline = "pcf_pipeline.cli:main"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: marks tests as slow (run with -m slow)"]
